+++ tests/Application/AbstractSignInWriteTest.php
@@ -33,6 +33,22 @@ abstract class AbstractSignInWriteTest extends AbstractApplicationTestCase
         ];
     }
 
+    public function testWriteUnauthorized(): void
+    {
+        $response = self::$staticApplicationClient->makeSignInPageWriteRequest('user@example.com', 'invalid');
+
+        self::assertSame(302, $response->getStatusCode());
+        self::assertSame('', $response->getHeaderLine('content-type'));
+        self::assertStringContainsString('/sign-in/', $response->getHeaderLine('location'));
+        self::assertSame('', $response->getBody()->getContents());
+
+        $responseCookieValue = $response->getHeaderLine('set-cookie');
+        if ('' !== $responseCookieValue) {
+            $responseCookie = Cookie::fromString($response->getHeaderLine('set-cookie'));
+            self::assertNotSame('token', $responseCookie->getName());
+        }
+    }
+
     public function testWriteSuccess(): void
     {
         $response = self::$staticApplicationClient->makeSignInPageWriteRequest('user@example.com', 'password');
